LayoutTests/imported/w3c/web-platform-tests/css/css-mixins/local-var-substitution-expected.txt


FAIL Fallback directly in result assert_equals: expected "PASS" but got "--f()"
FAIL Fallback via present, but invalid local assert_equals: expected "PASS" but got "--f()"
FAIL Fallback is locally resolved (result) assert_equals: expected "PASS" but got "--f()"
FAIL Fallback is locally resolved (local var) assert_equals: expected "PASS" but got "--f()"
